JOTM 18/09/2018 Funcionalidad recarga estación
Agrega funciones estaciones 
Queda pendiente reporte ,guardado en local, lisener y envió a rest

+++ 1.Layers/1.1Presentation/1.1.2Mobile/app/src/main/java/com/example/neotecknewts/sagasapp/Util/Lisener.java
@@ -153,9 +153,9 @@ public class Lisener{
                             cursor.getColumnIndex("IdTipoMedidorEntrada")));
                     recargaDTO.setIdTipoEvento(cursor.getInt(
                             cursor.getColumnIndex("IdTipoEvento")));
-                    recargaDTO.setP5000Salida(cursor.getDouble(
+                    recargaDTO.setP5000Salida(cursor.getInt(
                             cursor.getColumnIndex("P5000Salida")));
-                    recargaDTO.setP5000Entrada(cursor.getDouble(
+                    recargaDTO.setP5000Entrada(cursor.getInt(
                             cursor.getColumnIndex("P5000Entrada")));
 
                     String tipo = cursor.getString(
